+++ lib/src/languages/java/java11/java11_generator.dart
@@ -11,7 +11,7 @@ class ApolloCodeGeneratorJava11 extends ApolloCodeGenerator {
       [String indent = '', StringBuffer? s]) {
     s ??= StringBuffer();
 
-    var code = generateASTBlock(clazz);
+    var code = generateASTBlock(clazz, '', null, true, true);
 
     s.write('class ');
     s.write(clazz.name);
@@ -60,7 +60,7 @@ class ApolloCodeGeneratorJava11 extends ApolloCodeGenerator {
     s.write(') {\n');
     s.write(blockCode);
     s.write(indent);
-    s.write('}\n');
+    s.write('}\n\n');
 
     return blockCode;
   }